Split TextEditor class
Now there is also a TabManager class.
