+++ FileButton.py
@@ -10,12 +10,12 @@ class FileButton(ttk.Button):
             app.button_frame,
             style="File.TButton",
             text=displayed_name,
-            command=lambda: app.switch_tabs(raw_file)
+            command=lambda: app.tab_manager.switch_tabs(raw_file)
         )
         self.raw_file = raw_file
         self.pack(side=tk.LEFT, fill=tk.Y)
         self.config(width=len(displayed_name))
         self.bind(
-            '<Button-2>',  # Right click
-            lambda event: app.close_file(self.raw_file)
+            '<Button-2>',  # Right click for Mac / Wheel click for Windows
+            lambda event: app.tab_manager.close_file(self.raw_file)
         )
